% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap Model.R
\name{bootstrapped.extraction}
\alias{bootstrapped.extraction}
\title{Bootstrap model for the Stimson dyad ratios algorithm.}
\usage{
bootstrapped.extraction(data, reps = 1000, draw = 0.2, varname, output,
  print = FALSE)
}
\arguments{
\item{data}{Define the data}

\item{reps}{Define number of bootstrapped replications. Default is 500.}

\item{draw}{Define the proportion of items to be dropped in each bootstrap replication. Default is 0.1.}

\item{varname}{Define the variable name indicating the input series (as in extract function)}

\item{output}{Define the object name that extract function results are stores in.}

\item{print}{Logical. Define whether or not the function should display active count of replication progress in the console.}
}
\description{
Runs bootstrap model for repeated estimation of variable loading scores in the dyad-ratios algorithm.
}
\details{
This function runs the dyad-ratios bootstrap model, which takes the results of a single dyad-ratios estimation outcome and produces bootstrapped estimations of the variable loading scores.

The bootstrap model removes a pre-defined proportion of random variables for each estimation, extracts the variable loading scores, and averages them across all trials to create bootstrapped-mean estimations of item validity.

The model inherits all formula arguments from the extract function output. Note that the extract function must be as defined in this package code, not the original Extract.r.zip file hosted on Stimson's website.

Assigning the output to an object creates a list of five items, including a dataframe called 'Full Results' which contrains the bootstrapped mean loading score, single-run estimated loading score, and the difference between the two for each variable input. Larger figures indicate bigger over-estimation of the loading score in the single-run estimation (a lower bootstrapped mean).

It is strongly suggested (but not essential) for the speed of the bootstrap estimation that the options print and log in the extract function (to produce the results object) are set to FALSE.
}
