% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_dr_code.R
\name{call.dr.code}
\alias{call.dr.code}
\title{Call and Store Dyad-Ratios Functions}
\usage{
call.dr.code(filepath = "load_functions.R")
}
\arguments{
\item{filepath}{Character vector indicating the location and filename of the dyad-ratios functions. Defula is load_function.R}
}
\description{
Collects and assigns functions required to run the Stimson dyad-ratios model.
}
\details{
Simple function to source and assign all dyad-ratios model functions into the global environment.

Takes filepath argument, which by default looks in the same folder for load_functions.R.

Sigificant changes have been made to the extract function (non-calculation elements) which are documented in the main package documentation.
}
