% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{display}
\alias{display}
\title{Dyad-Ratios Functions for assignment into Global Environment}
\usage{
display(out, filename = NULL)
}
\description{
Storage for Stimson functions, with significant changes and bug fixes within the extract and aggregate functions.
}
\details{
No arguments or functions to call, assigned to environment using the call_dr_code command.
}
